YEL-40 [feat] 코드리뷰 반영
