--- src/main/java/com/yello/server/domain/question/service/YelloVoteService.java
@@ -1,12 +1,0 @@
-package com.yello.server.domain.question.service;
-
-import com.yello.server.domain.question.dto.response.YelloStartResponse;
-import com.yello.server.domain.question.dto.response.YelloVoteResponse;
-
-import java.util.List;
-
-public interface YelloVoteService {
-    List<YelloVoteResponse> findYelloVoteList(Long userId);
-
-    YelloStartResponse yelloStartCheck(Long userId);
-}
