+++ src/main/java/com/yello/server/global/common/util/TimeUtil.java
@@ -2,6 +2,7 @@ package com.yello.server.global.common.util;
 
 import java.time.Duration;
 import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 
 public class TimeUtil {
 
@@ -23,5 +24,10 @@ public class TimeUtil {
         }
 
         return (seconds) + "일 전";
+    }
+
+    public static String toDateFormattedString(LocalDateTime localDateTime) {
+        DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
+        return localDateTime.format(dateTimeFormatter).toString();
     }
 }
